Fix after merge and conflicts resolve + wording change basket to cart + route for myorders
